Moves WebShare component from core to themes
