+++ src/themes/default/components/theme/WebShare.vue
